Rewrite the test checking the POM generation in order to use TestProjectRule
